refactor(client): extract runBatchUpload into the pipeline module
